+++ src/Exception/ContextException.php
@@ -28,4 +28,15 @@ final class ContextException extends RuntimeException implements MoneyException
     {
         return new self('AutoContext only supports RoundingMode::Unnecessary.');
     }
+
+    /**
+     * @pure
+     */
+    public static function notFixedContext(string $method): self
+    {
+        return new self(sprintf(
+            '%s() does not support AutoContext; use toContext() to convert to a fixed context first.',
+            $method,
+        ));
+    }
 }